% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuliaObjectExt.R
\name{autowrap}
\alias{autowrap}
\title{Use automatic wrapper for julia type.}
\usage{
autowrap(type, fields = NULL, methods = c())
}
\arguments{
\item{type}{the julia type to wrap.}

\item{fields}{names of fields to be included in the wrapper.
If the value is NULL, then every julia fields will be included in the wrapper.}

\item{methods}{names of methods to be overloaded for the wrapper.}
}
\description{
\code{autowrap} tells `JuliaCall` to use automatic wrapper for julia type.
}
