% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installJulia.R
\name{.latest_lts_julia_version_from_versions}
\alias{.latest_lts_julia_version_from_versions}
\title{Get the latest Long Term Support (LTS) Julia version.}
\usage{
.latest_lts_julia_version_from_versions(versions)
}
\value{
A numeric_version object representing the latest LTS Julia version.
}
\description{
Get the latest Long Term Support (LTS) Julia version.
}
