% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installJulia.R
\name{.latest_stable_julia_version_from_versions}
\alias{.latest_stable_julia_version_from_versions}
\title{Get the latest stable Julia version.}
\usage{
.latest_stable_julia_version_from_versions(versions)
}
\value{
A numeric_version object representing the latest stable Julia version.
}
\description{
Get the latest stable Julia version.
}
