% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMarkdown.R
\name{eng_juliacall}
\alias{eng_juliacall}
\title{Julia language engine in R Markdown}
\usage{
eng_juliacall(options)
}
\arguments{
\item{options}{a list of chunk options}
}
\description{
Julia language engine in R Markdown
}
\examples{

knitr::knit_engines$set(julia = JuliaCall::eng_juliacall)

}
