% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installJulia.R
\name{install_julia}
\alias{install_julia}
\title{Install Julia.}
\usage{
install_julia(version = "latest", prefix = julia_default_install_dir())
}
\arguments{
\item{version}{The version of Julia to install (e.g. \code{"1.6.3"}).
Defaults to \code{"latest"}, which will install the most
recent stable release.}

\item{prefix}{the directory where Julia will be installed.
If not set, a default location will be determined by \code{rappdirs}
if it is installed, otherwise an error will be raised.}
}
\description{
Install Julia.
}
