% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{julia_assign}
\alias{julia_assign}
\title{Assign a value to a name in julia.}
\usage{
julia_assign(x, value)
}
\arguments{
\item{x}{a variable name, given as a character string.}

\item{value}{a value to be assigned to x, note that R value will be converted to
corresponding julia value automatically.}
}
\description{
\code{julia_assign} assigns a value to a name in julia with automatic type conversion.
}
\examples{

\donttest{ ## julia_setup is quite time consuming
  ## doing initialization and automatic installation of Julia if necessary
  julia_setup(installJulia = TRUE)
  julia_assign("x", 2)
  julia_assign("rsqrt", sqrt)
}

}
