% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{julia_call}
\alias{julia_call}
\title{Call julia functions.}
\usage{
julia_call(func_name, ..., need_return = TRUE)
}
\arguments{
\item{func_name}{the name of julia function you want to call.}

\item{...}{the unnamed arguments you want to pass to the julia function.}

\item{need_return}{whether you want the julia return value or not.}
}
\description{
\code{julia_call} calls julia functions with automatic type conversion.
}
\examples{

if (julia_check()) {
  \dontrun{ ## julia_setup is quite time consuming
  julia <- julia_setup()

  julia_call("sqrt", 2)
  }
}

}
