% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{julia_check}
\alias{julia_check}
\title{Check whether julia is available on the path.}
\usage{
julia_check()
}
\value{
whether julia is available on the path.
}
\description{
\code{julia_check} checks if julia is available on the path.
}
\examples{
julia_check()

}
