% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{julia_command}
\alias{julia_command}
\title{Evaluate string commands in julia.}
\usage{
julia_command(cmd)
}
\arguments{
\item{cmd}{the command string you want to evaluate in julia.}
}
\description{
\code{julia_command} evaluates string commands in julia
without returning the result.
If you need the result, maybe you could
try \code{julia_eval_string}.
}
\examples{

if (julia_check()) {
  julia <- julia_setup()

  julia_command("a = sqrt(2)")
}

}
