% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/console.R
\name{julia_console}
\alias{julia_console}
\title{Open julia console.}
\usage{
julia_console()
}
\description{
Open julia console.
}
\examples{
if (identical(Sys.getenv("AUTO_JULIA_INSTALL"), "true")) { ## julia_setup is quite time consuming
  julia_console()
}

}
