% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{julia_do.call}
\alias{julia_do.call}
\title{\code{do.call} for julia.}
\usage{
julia_do.call(func_name, arg_list, need_return = TRUE)
}
\arguments{
\item{func_name}{the name of julia function you want to call.}

\item{arg_list}{the unnamed list of the arguments you want to pass to the julia function.}

\item{need_return}{whether you want the julia to return value or not.}
}
\description{
\code{julia_do.call} is the \code{do.call} for julia with automatic type conversion.
}
\examples{

if (julia_check()) {
  \dontrun{ ## julia_setup is quite time consuming
  julia <- julia_setup()

  julia_do.call("sqrt", list(2))
  }
}

}
