% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{julia_eval_string}
\alias{julia_eval_string}
\title{Evalulate string command in julia and get the return value.}
\usage{
julia_eval_string(cmd)
}
\arguments{
\item{cmd}{the command string you want to evaluate in julia.}
}
\value{
the R object automatically converted from julia object.
}
\description{
\code{julia_eval_string} evaluates string command in julia and
returns the result as an R object.
If you don't need the result, maybe you could
consider using \code{julia_command}.
}
\examples{

if (julia_check()) {
  julia <- julia_setup()

  julia_eval_string("sqrt(2)")
}

}
