% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{julia_exists}
\alias{julia_exists}
\title{Check whether a julia object with the given name exists or not.}
\usage{
julia_exists(name)
}
\arguments{
\item{name}{the name of julia object you want to check.}
}
\description{
\code{julia_exists} returns whether a julia object with the given name exists or not.
}
\examples{

\dontrun{ ## julia_setup is quite time consuming
  julia <- julia_setup()

  julia_exists("sqrt")
}

}
