% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installJulia.R
\name{julia_latest_version}
\alias{julia_latest_version}
\title{Get the latest stable Julia version.}
\usage{
julia_latest_version()
}
\value{
A numeric_version object representing the latest stable Julia version.
}
\description{
Get the latest stable Julia version.
}
