% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installJulia.R
\name{julia_lts_version}
\alias{julia_lts_version}
\title{Get the latest Long Term Support (LTS) Julia version.}
\usage{
julia_lts_version()
}
\value{
A numeric_version object representing the latest LTS Julia version.
}
\description{
Get the latest Long Term Support (LTS) Julia version.
}
