% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMarkdown.R
\name{julia_markdown_setup}
\alias{julia_markdown_setup}
\title{Do setup for JuliaCall in RMarkdown documents and notebooks.}
\usage{
julia_markdown_setup(..., notebook = FALSE)
}
\arguments{
\item{...}{The same arguments accepted by `julia_setup`.}

\item{notebook}{whether it is in RStudio notebook environment or not.}
}
\description{
\code{julia_markdown_setup} does the initial setup for JuliaCall in RMarkdown document and RStudio notebooks.
  The function should be invoked automatically most of the case.
  It can also be called explicitly in RMarkdown documents or notebooks.
}
