% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMarkdown.R
\name{julia_notebook_setup}
\alias{julia_notebook_setup}
\title{Do setup for julia chunks in RMarkdown notebooks.}
\usage{
julia_notebook_setup(...)
}
\arguments{
\item{...}{The same arguments accepted by `julia_setup`.}
}
\description{
\code{julia_notebook_setup} does the initial setup for julia chunks in RMarkdown notebooks.
  The function must be called to deal with the
  "don't know how to handle julia engine output" problem in notebooks.
}
