% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{julia_source}
\alias{julia_source}
\title{Source a julia source file.}
\usage{
julia_source(file_name)
}
\arguments{
\item{file_name}{the name of julia source file.}
}
\description{
\code{julia_source} sources a julia source file.
}
