% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{juliacall_dependency}
\alias{juliacall_dependency}
\title{Install and Setup for Julia package dependency of JuliaCall.}
\usage{
juliacall_dependency()
}
\description{
\code{juliacall_dependency} install julia package dependencies for JuliaCall.
}
