% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuliaObject.R
\name{JuliaObject}
\alias{JuliaObject}
\title{Convert an R Object to Julia Object.}
\usage{
JuliaObject(x)
}
\arguments{
\item{x}{the R object you want to convert to julia object.}
}
\value{
a S3 object with class JuliaObject
which contains a reference to the corresponding julia object.
}
\description{
\code{JuliaObject} converts an R object to julia object
  and returns a reference of the corresponding julia object.
}
\examples{

\dontrun{ ## julia_setup is quite time consuming
  a <- JuliaObject(1)
}

}
