% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuliaObject.R
\name{JuliaObjectFields}
\alias{JuliaObjectFields}
\alias{fields}
\alias{fields.JuliaObject}
\alias{field}
\alias{field.JuliaObject}
\alias{field<-}
\alias{field<-.JuliaObject}
\title{JuliaObject Fields.}
\usage{
fields(object)

\method{fields}{JuliaObject}(object)

field(object, name)

\method{field}{JuliaObject}(object, name)

field(object, name) <- value

\method{field}{JuliaObject}(object, name) <- value
}
\arguments{
\item{object}{the JuliaObject.}

\item{name}{a character string specifying the fields to be accessed or set.}

\item{value}{the new value of the field of the JuliaObject.}
}
\description{
Get the field names, get or set certain fields of an JuliaObject.
}
