% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuliaObject.R
\name{print.JuliaObject}
\alias{print.JuliaObject}
\title{Print Julia Object.}
\usage{
\method{print}{JuliaObject}(x, ...)
}
\arguments{
\item{x}{the Julia object you want to print.}

\item{...}{further arguments to be passed to or from other methods.
They are ignored in this function.}
}
\description{
\code{print.JuliaObject} is the method of the generic print function
for JuliaObject.
}
