% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_generic.R
\name{rep.JuliaObject}
\alias{rep.JuliaObject}
\title{#' @export
unlist.JuliaObject <- function(x, recursive = TRUE, use.names = TRUE){
    tryCatch(julia_call("JuliaCall.vcat", x),
             warn = function(e){},
             error = function(e) x
    )
}}
\usage{
\method{rep}{JuliaObject}(x, times, ...)
}
\description{
#' @export
unlist.JuliaObject <- function(x, recursive = TRUE, use.names = TRUE){
    tryCatch(julia_call("JuliaCall.vcat", x),
             warn = function(e){},
             error = function(e) x
    )
}
}
