% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\docType{methods}
\name{S4JuliaObjectGeneric}
\alias{S4JuliaObjectGeneric}
\alias{Compare,JuliaObject,ANY-method}
\alias{Compare,ANY,JuliaObject-method}
\alias{Arith,JuliaObject,ANY-method}
\alias{Arith,ANY,JuliaObject-method}
\alias{\%\%,JuliaObject,ANY-method}
\alias{\%\%,ANY,JuliaObject-method}
\alias{\%/\%,JuliaObject,ANY-method}
\alias{\%/\%,ANY,JuliaObject-method}
\alias{Math,JuliaObject-method}
\alias{Math2,JuliaObject-method}
\alias{Summary,JuliaObject-method}
\title{S4 Group Generic Functions for JuliaObject.}
\usage{
\S4method{Compare}{JuliaObject,ANY}(e1, e2)

\S4method{Compare}{ANY,JuliaObject}(e1, e2)

\S4method{Arith}{JuliaObject,ANY}(e1, e2)

\S4method{Arith}{ANY,JuliaObject}(e1, e2)

\S4method{\%\%}{JuliaObject,ANY}(e1, e2)

\S4method{\%\%}{ANY,JuliaObject}(e1, e2)

\S4method{\%/\%}{JuliaObject,ANY}(e1, e2)

\S4method{\%/\%}{ANY,JuliaObject}(e1, e2)

\S4method{Math}{JuliaObject}(x)

\S4method{Math2}{JuliaObject}(x, digits)

\S4method{Summary}{JuliaObject}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x, e1, e2}{objects.}

\item{digits}{number of digits to be used in round or signif.}

\item{...}{further arguments passed to or from methods.}

\item{na.rm}{logical: should missing values be removed?}
}
\description{
S4 Group Generic Functions for JuliaObject, which includes common mathematical
  functions and operators.
}
