% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuliaObject.R
\docType{methods}
\name{show,JuliaObject-method}
\alias{show,JuliaObject-method}
\title{Show JuliaObject.}
\usage{
\S4method{show}{JuliaObject}(object)
}
\arguments{
\item{object}{the JuliaObject you want to show.}
}
\description{
S4 method to show JuliaObject.
}
