% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{julia_eval_string}
\alias{julia_eval_string}
\title{Evaluate string commands in julia and get the result.}
\usage{
julia_eval_string(cmd)
}
\arguments{
\item{cmd}{the command string you want to evaluate in julia.}
}
\description{
Deprecated in favor of julia_eval.
}
