% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{julia_eval_string}
\alias{julia_eval_string}
\title{Evaluate string commands in julia and get the result.}
\usage{
julia_eval_string(cmd)
}
\arguments{
\item{cmd}{the command string you want to evaluate in julia.}
}
\value{
the R object automatically converted from julia object.
}
\description{
\code{julia_eval_string} evaluates string commands in julia and
returns the result (automatically converted to an R object).
If you don't need the result, maybe you could
try \code{julia_command}.
}
\examples{

\dontrun{ ## julia_setup is quite time consuming
  julia <- julia_setup()

  julia_eval_string("sqrt(2)")
}

}
