% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{julia_help}
\alias{julia_help}
\title{Get help for a julia function.}
\usage{
julia_help(fname)
}
\arguments{
\item{fname}{the name of julia function you want to get help with.}
}
\description{
\code{julia_help} outputs the documentation of a julia function.
}
\examples{

\dontrun{ ## julia_setup is quite time consuming
  julia_help("sqrt")
}

}
