% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{julia_setup}
\alias{julia_setup}
\title{Do initial setup for the JuliaCall package.}
\usage{
julia_setup(verbose = FALSE)
}
\arguments{
\item{verbose}{whether print out detailed information about construction of julia interface}
}
\value{
The julia interface, which is an environment with the necessary methods
  like cmd, source and things like that to communicate with julia.
}
\description{
\code{julia_setup} does the initial setup for the JuliaCall package.
}
\examples{
julia <- julia_setup()

## Different ways for calculating sqrt(2)

julia$command("a = sqrt(2)"); julia$eval_string("a")
julia$eval_string("sqrt(2)")
julia$call("sqrt", 2)
julia$eval_string("sqrt")(2)

## You can use `julia$exists` as `exists` in R to test
## whether a function or name exists in Julia or not

julia$exists("sqrt")
julia$exists("c")

## You can use `julia$help` to get help for Julia functions

julia$help("sqrt")

## Functions related to Julia packages

julia$install_package("Optim")
julia$install_package_if_needed("Optim")
julia$installed_package("Optim")
julia$using("Optim") ## Same as julia$library("Optim")

}
