% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{julia_setup}
\alias{julia_setup}
\title{Do initial setup for the JuliaCall package.}
\usage{
julia_setup(JULIA_HOME = NULL, verbose = TRUE, force = FALSE,
  useRCall = TRUE)
}
\arguments{
\item{JULIA_HOME}{the file folder which contains julia binary,
if not set, JuliaCall will look at the global option JULIA_HOME,
if the global option is not set,
JuliaCall will then look at the environmental variable JULIA_HOME,
if still not found, JuliaCall will try to use
the julia in path.}

\item{verbose}{whether to print out detailed information
about \code{julia_setup}.}

\item{force}{whether to force julia_setup to execute again.}

\item{useRCall}{whether or not you want to use RCall.jl in julia,
which is an amazing package to access R in julia.}
}
\value{
The julia interface, which is an environment with the necessary methods
  like command, source and things like that to communicate with julia.
}
\description{
\code{julia_setup} does the initial setup for the JuliaCall package.
}
\examples{

\dontrun{ ## julia_setup is quite time consuming
  julia <- julia_setup()
}

}
