% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{julia_setup}
\alias{julia_setup}
\title{Do initial setup for the JuliaCall package.}
\usage{
julia_setup(verbose = TRUE, force = FALSE)
}
\arguments{
\item{verbose}{whether to print out detailed information
about \code{julia_setup}.}

\item{force}{whether to force julia_setup to execute again.}
}
\value{
The julia interface, which is an environment with the necessary methods
  like cmd, source and things like that to communicate with julia.
}
\description{
\code{julia_setup} does the initial setup for the JuliaCall package.
}
\examples{

if (julia_check()) {
  julia <- julia_setup()
}

}
