% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{julia_setup}
\alias{julia_setup}
\title{Do initial setup for the JuliaCall package.}
\usage{
julia_setup()
}
\value{
The julia interface, which is an environment with the necessary methods
  like cmd, source and things like that to communicate with julia.
}
\description{
\code{julia_setup} does the initial setup for the JuliaCall package.
}
\examples{
julia <- julia_setup()
julia$cmd("println(sqrt(2))")
julia$eval_string("sqrt(2)")
julia$call("sqrt", 2)
julia$eval_string("sqrt")(2)
julia$exists("sqrt")
julia$exists("c")

}
