% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JuliaObject.R
\name{JuliaObject}
\alias{JuliaObject}
\title{Convert an R Object to Julia Object.}
\usage{
JuliaObject(x)
}
\arguments{
\item{x}{the R object you want to convert to julia object.}
}
\value{
an environment of class JuliaObject,
which contains an id corresponding to the actual julia object.
}
\description{
\code{JuliaObject} converts an R object to julia object
  and returns a reference of the corresponding julia object.
}
\examples{

if (interactive()) { ## julia_setup is quite time consuming
  ## doing initialization and automatic installation of Julia if necessary
  julia_setup(installJulia = TRUE)
  a <- JuliaObject(1)
}

}
