% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\docType{methods}
\name{S4JuliaObjectGeneric}
\alias{S4JuliaObjectGeneric}
\alias{Compare,JuliaObject,ANY-method}
\alias{Compare,ANY,JuliaObject-method}
\alias{Compare,JuliaObject,JuliaObject-method}
\alias{Arith,JuliaObject,ANY-method}
\alias{Arith,ANY,JuliaObject-method}
\alias{Arith,JuliaObject,JuliaObject-method}
\alias{\%\%,JuliaObject,ANY-method}
\alias{\%\%,ANY,JuliaObject-method}
\alias{\%\%,JuliaObject,JuliaObject-method}
\alias{\%/\%,JuliaObject,ANY-method}
\alias{\%/\%,ANY,JuliaObject-method}
\alias{\%/\%,JuliaObject,JuliaObject-method}
\alias{Math,JuliaObject-method}
\alias{ceiling,JuliaObject-method}
\alias{cummax,JuliaObject-method}
\alias{cummin,JuliaObject-method}
\alias{cumsum,JuliaObject-method}
\alias{cumprod,JuliaObject-method}
\alias{tanpi,JuliaObject-method}
\alias{round,JuliaObject-method}
\alias{signif,JuliaObject-method}
\alias{Summary,JuliaObject-method}
\title{S4 Group Generic Functions for JuliaObject.}
\usage{
\S4method{Compare}{JuliaObject,ANY}(e1, e2)

\S4method{Compare}{ANY,JuliaObject}(e1, e2)

\S4method{Compare}{JuliaObject,JuliaObject}(e1, e2)

\S4method{Arith}{JuliaObject,ANY}(e1, e2)

\S4method{Arith}{ANY,JuliaObject}(e1, e2)

\S4method{Arith}{JuliaObject,JuliaObject}(e1, e2)

\S4method{\%\%}{JuliaObject,ANY}(e1, e2)

\S4method{\%\%}{ANY,JuliaObject}(e1, e2)

\S4method{\%\%}{JuliaObject,JuliaObject}(e1, e2)

\S4method{\%/\%}{JuliaObject,ANY}(e1, e2)

\S4method{\%/\%}{ANY,JuliaObject}(e1, e2)

\S4method{\%/\%}{JuliaObject,JuliaObject}(e1, e2)

\S4method{Math}{JuliaObject}(x)

\S4method{ceiling}{JuliaObject}(x)

\S4method{cummax}{JuliaObject}(x)

\S4method{cummin}{JuliaObject}(x)

\S4method{cumsum}{JuliaObject}(x)

\S4method{cumprod}{JuliaObject}(x)

\S4method{tanpi}{JuliaObject}(x)

\S4method{round}{JuliaObject}(x, digits = 0)

\S4method{signif}{JuliaObject}(x, digits = 6)

\S4method{Summary}{JuliaObject}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x, e1, e2}{objects.}

\item{digits}{number of digits to be used in round or signif.}

\item{...}{further arguments passed to or from methods.}

\item{na.rm}{logical: should missing values be removed?}
}
\description{
S4 Group Generic Functions for JuliaObject, which includes common mathematical
  functions and operators.
}
