% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatch.R
\docType{methods}
\name{S4JuliaObjectGeneric}
\alias{S4JuliaObjectGeneric}
\alias{Compare,JuliaObject,ANY-method}
\alias{Compare,ANY,JuliaObject-method}
\alias{Compare,JuliaObject,JuliaObject-method}
\alias{Arith,JuliaObject,ANY-method}
\alias{Arith,ANY,JuliaObject-method}
\alias{Arith,JuliaObject,JuliaObject-method}
\alias{\%\%,JuliaObject,ANY-method}
\alias{\%\%,ANY,JuliaObject-method}
\alias{\%\%,JuliaObject,JuliaObject-method}
\alias{\%/\%,JuliaObject,ANY-method}
\alias{\%/\%,ANY,JuliaObject-method}
\alias{\%/\%,JuliaObject,JuliaObject-method}
\alias{Logic,JuliaObject,ANY-method}
\alias{Logic,ANY,JuliaObject-method}
\alias{Logic,JuliaObject,JuliaObject-method}
\title{S4 Group Generic Functions for JuliaObject.}
\usage{
\S4method{Compare}{JuliaObject,ANY}(e1, e2)

\S4method{Compare}{ANY,JuliaObject}(e1, e2)

\S4method{Compare}{JuliaObject,JuliaObject}(e1, e2)

\S4method{Arith}{JuliaObject,ANY}(e1, e2)

\S4method{Arith}{ANY,JuliaObject}(e1, e2)

\S4method{Arith}{JuliaObject,JuliaObject}(e1, e2)

\S4method{\%\%}{JuliaObject,ANY}(e1, e2)

\S4method{\%\%}{ANY,JuliaObject}(e1, e2)

\S4method{\%\%}{JuliaObject,JuliaObject}(e1, e2)

\S4method{\%/\%}{JuliaObject,ANY}(e1, e2)

\S4method{\%/\%}{ANY,JuliaObject}(e1, e2)

\S4method{\%/\%}{JuliaObject,JuliaObject}(e1, e2)

\S4method{Logic}{JuliaObject,ANY}(e1, e2)

\S4method{Logic}{ANY,JuliaObject}(e1, e2)

\S4method{Logic}{JuliaObject,JuliaObject}(e1, e2)
}
\arguments{
\item{e1, e2}{objects.}
}
\description{
S4 Group Generic Functions for JuliaObject, which includes common mathematical
  functions and operators.
}
